New Features:
    1. Remove first color from color picker
    2. Also check userid for org users if their oid assignment s not
found in
    .creativeworx file.
    3. If user is not having permission to write into .creativeworx
file,
    display appropriate message.
    4. Make color Box:3 x 8.
    5. While editing and displaying the projects, make the Unassigned
    color(#888888) as new default color(#999999)

+++ jsx/cwprojectFileUtils.jsx
@@ -177,6 +177,9 @@
 			parentFolder = myFile.parent;
 			var file = this.checkIfFileExists(parentFolder);
 			var file = new File(parentFolder+"/.creativeworxproject");
+			if(file.readonly){
+				return "PRMDND"
+			}
 			this.writeFile(file, newPair);
 			return "true";
 		}
